+++ PokerNowKit/Models/Seat.swift
@@ -13,6 +13,6 @@ struct Seat {
     var summary: String = ""
     var preFlopBet: Bool = false
     var showedHand: String?
-    var stack: Int?
+    var stack: Double?
     var number: Int = 0
 }
